+++ src/Clickup/Parsers/TasksFileParser.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace PlacetoPay\AppVersion\Clickup\Parsers;
+
+use PlacetoPay\AppVersion\Exceptions\ChangelogException;
+use PlacetoPay\AppVersion\Helpers\ChangelogLastChanges;
+
+class TasksFileParser
+{
+    public const REGEX_CLICKUP_IDENTIFIER = '/\(https:\/\/app\.clickup\.com\/t(?:\/(?<team>\d+))?\/(?<id>[\w-]+)\)/';
+    private ChangelogLastChanges $changelog;
+
+    public function __construct(ChangelogLastChanges $changelog)
+    {
+        $this->changelog = $changelog;
+    }
+
+    /**
+     * @throws ChangelogException
+     */
+    public function tasksData(string $changelogFileName): ?array
+    {
+        $this->changelog->read($changelogFileName);
+
+        $content = $this->changelog->content();
+        if (empty($content)) {
+            return null;
+        }
+
+        $tasks = $this->extractTasks($content);
+        if (empty($tasks)) {
+            return null;
+        }
+
+        return [
+            'version' => $this->changelog->version(),
+            'tasks' => $tasks,
+        ];
+    }
+
+    private function extractTasks(array $changes): array
+    {
+        $tasks = [];
+
+        foreach ($changes as $change) {
+            if (preg_match(self::REGEX_CLICKUP_IDENTIFIER, $change, $match)) {
+                $tasks[] = [
+                    'id' => $match['id'],
+                    'team' => $match['team'] ?: null,
+                ];
+            }
+        }
+
+        return $tasks;
+    }
+}