+++ changelog.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 3.1.1 (2025-05-22)
-  - Allow use env custom to define env when to execute the command create-deploy
+## 3.3.0 (2025-09-03)
+- Allow use env custom to define env when to execute the command create-deploy
+
+## 3.2.0 (2025-07-02)
+- Add support for posting a comment to ClickUp using the `app-version:notify-clickup` command.
 
 ## 3.1.0 (2025-05-09)
 - Adds support for Laravel 12